# docs/swagger.yaml
basePath: /
definitions:
  handlers.AuthResponse:
    properties:
      expires_at:
        example: 1701734400
        type: integer
      rol:
        example: Alumno
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user_id:
        example: 1
        type: integer
      username:
        example: jperez2025
        type: string
    type: object
  handlers.LoginRequest:
    properties:
      password:
        example: Segura123#
        type: string
      username:
        example: jperez2025
        type: string
    type: object
  handlers.RegisterRequest:
    properties:
      email:
        example: juan.perez@universidad.edu.mx
        type: string
      first_name:
        example: Juan
        type: string
      last_name:
        example: Pérez García
        type: string
      matricula:
        example: "20250001"
        type: string
      password:
        example: Segura123#
        type: string
      rol:
        example: Alumno
        type: string
      username:
        example: jperez2025
        type: string
    type: object
...
  models.Cuatrimestre:
    description: Modelo de Cuatrimestre utilizado en la API y sincronizado como subcategoría
      en Moodle.
    properties:
      descripcion:
        example: Cuatrimestre correspondiente al periodo enero-abril 2025
        type: string
      id_externo:
        example: CUATR-2025-01
        type: string
      id_moodle:
        example: 5678
        type: integer
      nombre:
        example: Primer Cuatrimestre 2025
        type: string
      programa_estudio_id:
        description: Campo de la Clave Foránea
        example: 3
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API para gestión de Programa de Estudio, Cuatrimestres, Asignaturas,
    Usuarios y Grupos con sincronización a Moodle.
  title: Control Escolar API
  version: "2.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Autentica un usuario con username y password, devuelve un token
        JWT
      parameters:
      - description: Credenciales de inicio de sesión
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Autenticación exitosa con token JWT
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "400":
          description: Datos inválidos
          schema:
            type: string
        "401":
          description: Credenciales incorrectas
          schema:
            type: string
        "500":
          description: Error interno del servidor
          schema:
            type: string
      summary: Iniciar sesión
      tags:
      - Autenticación
  /auth/register:
    post:
      consumes:
      - application/json
      description: Registra un nuevo usuario (Docente o Alumno) con validación de
        contraseña y hash seguro
      parameters:
      - description: Datos del usuario para registro
        in: body
        name: usuario
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Usuario registrado exitosamente con token JWT
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "400":
          description: Datos inválidos o contraseña débil
          schema:
            type: string
        "409":
          description: Usuario ya existe
          schema:
            type: string
        "500":
          description: Error interno del servidor
          schema:
            type: string
      summary: Registrar un nuevo usuario
      tags:
      - Autenticación
  /cuatrimestre/:
    get:
      description: Obtiene todos los cuatrimestres
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Cuatrimestre'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Listar Cuatrimestres
      tags:
      - cuatrimestre
    post:
      consumes:
      - application/json
      description: Crea un cuatrimestre local
      parameters:
      - description: Datos del cuatrimestre
        in: body
        name: cuatrimestre
        required: true
        schema:
          $ref: '#/definitions/models.Cuatrimestre'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Cuatrimestre'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Crear Cuatrimestre
      tags:
      - cuatrimestre
  /cuatrimestre/{id}/:
    delete:
      description: Elimina un cuatrimestre por ID
      parameters:
      - description: ID del cuatrimestre
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Eliminar Cuatrimestre
      tags:
      - cuatrimestre
    get:
      description: Obtiene un cuatrimestre por ID
      parameters:
      - description: ID del cuatrimestre
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Cuatrimestre'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Obtener Cuatrimestre
      tags:
      - cuatrimestre
    put:
      consumes:
      - application/json
      description: Actualiza un cuatrimestre por ID
      parameters:
      - description: ID del cuatrimestre
        in: path
        name: id
        required: true
        type: integer
      - description: Datos del cuatrimestre
        in: body
        name: cuatrimestre
        required: true
        schema:
          $ref: '#/definitions/models.Cuatrimestre'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Cuatrimestre'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Actualizar Cuatrimestre
      tags:
      - cuatrimestre
  /cuatrimestre/bulk-sync:
    post:
      description: Sincroniza todos los cuatrimestres que no tienen ID_Moodle a Moodle
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Sincronización masiva de Cuatrimestres
      tags:
      - cuatrimestre
  /cuatrimestre/sync/{id}:
    post:
      description: Inicia la sincronización del cuatrimestre a Moodle
      parameters:
      - description: ID del cuatrimestre
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Sincronizar Cuatrimestre
      tags:
      - cuatrimestre
...
swagger: "2.0"